+++ packages/account-tree-controller/src/AccountTreeController.ts
@@ -29,6 +29,7 @@ import {
   isAccountGroupNameUniqueFromWallet,
   MAX_SORT_ORDER,
 } from './group';
+import { projectLogger as log } from './logger';
 import type { Rule } from './rule';
 import { EntropyRule } from './rules/entropy';
 import { KeyringRule } from './rules/keyring';
@@ -256,6 +257,8 @@ export class AccountTreeController extends BaseController<
       return;
     }
 
+    log('Initializing...');
+
     const wallets: AccountTreeControllerState['accountTree']['wallets'] = {};
 
     // Clear mappings for fresh rebuild.
@@ -267,8 +270,23 @@ export class AccountTreeController extends BaseController<
     const previousSelectedAccountGroup =
       this.state.accountTree.selectedAccountGroup;
 
+    // There's no guarantee that accounts would be sorted by their import time
+    // with `listMultichainAccounts`. We have to sort them here before constructing
+    // the tree.
+    //
+    // Because of the alignment mecanism, some accounts from the same group might not
+    // have been imported at the same time, but at least of them should have been
+    // imported at the right time, thus, inserting the group at the proper place too.
+    //
+    // Lastly, if one day we allow to have "gaps" in between groups, then this `sort`
+    // won't be enough and we would have to use group properties instead (like group
+    // index or maybe introduce a `importTime` at group level).
+    const accounts = this.#listAccounts().sort(
+      (a, b) => a.metadata.importTime - b.metadata.importTime,
+    );
+
     // For now, we always re-compute all wallets, we do not re-use the existing state.
-    for (const account of this.#listAccounts()) {
+    for (const account of accounts) {
       this.#insert(wallets, account);
     }
 
@@ -325,6 +343,9 @@ export class AccountTreeController extends BaseController<
       previousSelectedAccountGroup !==
       this.state.accountTree.selectedAccountGroup
     ) {
+      log(
+        `Selected (initial) group is: [${this.state.accountTree.selectedAccountGroup}]`,
+      );
       this.messagingSystem.publish(
         `${controllerName}:selectedAccountGroupChange`,
         this.state.accountTree.selectedAccountGroup,
@@ -332,6 +353,7 @@ export class AccountTreeController extends BaseController<
       );
     }
 
+    log('Initialized!');
     this.#initialized = true;
   }
 
@@ -342,6 +364,7 @@ export class AccountTreeController extends BaseController<
    * cleared state. Use this when you need to force a full re-init even if already initialized.
    */
   reinit() {
+    log('Re-initializing...');
     this.#initialized = false;
     this.init();
   }
@@ -407,6 +430,7 @@ export class AccountTreeController extends BaseController<
         wallet.metadata.name =
           this.#getKeyringRule().getDefaultAccountWalletName(wallet);
       }
+      log(`[${wallet.id}] Set default name to: "${wallet.metadata.name}"`);
     }
   }
 
@@ -540,6 +564,7 @@ export class AccountTreeController extends BaseController<
 
       state.accountTree.wallets[walletId].groups[groupId].metadata.name =
         proposedName;
+      log(`[${group.id}] Set default name to: "${group.metadata.name}"`);
 
       // Persist the generated name to ensure consistency
       state.accountGroupsMetadata[groupId] ??= {};
@@ -810,6 +835,7 @@ export class AccountTreeController extends BaseController<
     const walletId = result.wallet.id;
     let wallet = wallets[walletId];
     if (!wallet) {
+      log(`[${walletId}] Added as new wallet`);
       wallets[walletId] = {
         ...result.wallet,
         status: 'ready',
@@ -835,6 +861,7 @@ export class AccountTreeController extends BaseController<
     const sortOrder = ACCOUNT_TYPE_TO_SORT_ORDER[type];
 
     if (!group) {
+      log(`[${walletId}] Add new group: [${groupId}]`);
       wallet.groups[groupId] = {
         ...result.group,
         // Type-wise, we are guaranteed to always have at least 1 account.
@@ -879,6 +906,9 @@ export class AccountTreeController extends BaseController<
         },
       );
     }
+    log(
+      `[${groupId}] Add new account: { id: "${account.id}", type: "${account.type}", address: "${account.address}"`,
+    );
 
     // Update the reverse mapping for this account.
     this.#accountIdToContext.set(account.id, {
@@ -971,6 +1001,11 @@ export class AccountTreeController extends BaseController<
     this.update((state) => {
       state.accountTree.selectedAccountGroup = groupId;
     });
+
+    log(
+      `Selected group is now: [${this.state.accountTree.selectedAccountGroup}]`,
+    );
+
     this.messagingSystem.publish(
       `${controllerName}:selectedAccountGroupChange`,
       groupId,
@@ -1212,6 +1247,10 @@ export class AccountTreeController extends BaseController<
       this.#assertAccountGroupNameIsUnique(groupId, finalName);
     }
 
+    log(
+      `[${groupId}] Set new name to: "${finalName}" (auto handle conflict: ${autoHandleConflict})`,
+    );
+
     this.update((state) => {
       /* istanbul ignore next */
       if (!state.accountGroupsMetadata[groupId]) {
@@ -1356,6 +1395,8 @@ export class AccountTreeController extends BaseController<
    * Also clears the backup and sync service state.
    */
   clearState(): void {
+    log('Clearing state');
+
     this.update(() => {
       return {
         ...getDefaultAccountTreeControllerState(),